refactor: adjust tanstack queries parameters
