+++ project/components/ui/modals/create-project-modal.tsx
@@ -1,4 +1,4 @@
-import { useCreateProject } from "@/hooks/useProjects";
+import { useCreateProject } from "@/hooks/use-projects";
 import { useState } from "react";
 import Button from "../buttons/button";
 import Input from "../input-fields/input";
